Allow skipping deletion with `delete: false` when delete_raw plugin is loaded
The delete_raw plugin will get deprecated after the new derivatives
plugin, as it produces unwanted side effects. In the meanwhile, we add a
:delete option for skipping deletion for the particular upload.

  uploader.upload(file, delete: false)

This will be needed in the derivation_endpoint plugin.

+++ lib/shrine/plugins/delete_raw.rb
@@ -16,7 +16,7 @@ class Shrine
         # Deletes the file that was uploaded, unless it's an UploadedFile.
         def copy(io, context)
           super
-          if io.respond_to?(:path) && io.path && delete_raw?
+          if io.respond_to?(:path) && io.path && delete_raw? && context[:delete] != false
             begin
               File.delete(io.path)
             rescue Errno::ENOENT
